Fallback to apt update
Fallback to using the traditional apt update && apt upgrade if apt isn't
at least version 3.0.

I don't know when the magic -U was introduced, so I'll assume it was
available in 3.0 because the commit introducing -U is dated 2023-04-11
and apt 3.0.x was released 2025-04-04, so theoretically that's far
enough in the future.